chore: split shape-like funcs off verbs/mod
